backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
google-genai==1.14.0
python-dotenv==1.1.0
python-multipart==0.0.20
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
